pyproject.toml
[project]
name = "stackr"
version = "0.1.0"
description = "A declarative, composable homelab deployment tool"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "httpx>=0.27",
]
tui = [
    "textual>=0.60",
]
web = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
]

[project.scripts]
stackr = "stackr.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true

[[tool.mypy.overrides]]
module = "stackr.tui"
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = ["stackr.web", "stackr.web.*"]
ignore_missing_imports = true
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "asyncio: run test as an async coroutine (requires pytest-asyncio)",
]
